+++ HISTORY.rst
@@ -4,10 +4,10 @@ History
 0.6.0 (TBD)
 -----------
 
+* Added support for dot notation and version 1.1.2 of the spec (issue #99). [rbp]
 * Bugfix: falsey values now coerced to strings using str().
-* Bugfix: issue #113: lambda return values for sections no longer pushed
-  onto context stack.
-* Bugfix: issue #114: lists of lambdas for sections were not rendered.
+* Bugfix: lambda return values for sections no longer pushed onto context stack (issue #113).
+* Bugfix: lists of lambdas for sections were not rendered (issue #114).
 
 0.5.1 (2012-04-24)
 ------------------